ECDR-120 - updated with feedback from review, and also moved collapsed the util packages and classes down to a single package.
